# Conflicts:
#	Online-Messaging-Backend/src/_tests_/DAO.spec.ts
#	package-lock.json

+++ Online-Messaging-Backend/src/routes/notifications/notifications.ts
@@ -5,6 +5,7 @@ import aws from "aws-sdk";
 import { HTTPResponseAndToken, JwtVerificationService } from "../../shared/jwt-verification-service";
 import { NotificationObject, NotificationsDAO } from "./NotificationsDAO";
 import { uuid } from "uuidv4";
+import { sanitizeInput } from "../../index";
 
 const PATH_POST_NEW_NOTIFICATION: string = "/";
 const PATH_DELETE_NOTIFICATION: string = "/:notificationId/insertedTime/:insertedTime";
@@ -30,12 +31,12 @@ router.post(PATH_POST_NEW_NOTIFICATION, (req, res) => {
             const notificationsDAO: NotificationsDAO = new NotificationsDAO(docClient);
 
             let newNotification: NotificationObject = {
-                channelId: req.body.channelId,
-                channelName: req.body.channelName,
-                channelType: req.body.channelType,
-                username: req.body.username,
-                message: req.body.message,
-                type: req.body.type,
+                channelId: sanitizeInput(req.body.channelId),
+                channelName: sanitizeInput(req.body.channelName),
+                channelType: sanitizeInput(req.body.channelType),
+                username: sanitizeInput(req.body.username),
+                message: sanitizeInput(req.body.message),
+                type: sanitizeInput(req.body.type),
                 fromFriend: req.body.fromFriend,
                 notificationId: uuid(),
                 insertedTime: Date.now()
@@ -117,9 +118,12 @@ router.delete(PATH_DELETE_ALL_MESSAGE_NOTIFICATIONS_FOR_USER_FOR_CHANNEL, (req, 
                 .getAllNotificationsForChannelAtUsername(req.params.channelId, req.params.username)
                 .then((data: Array<NotificationObject>) => {
                     for (let item of data) {
-                        notificationsDAO.deleteNotification(item.notificationId, item.insertedTime);
-                        res.status(200).send({ status: 200, message: "Message notifications deleted successfully" });
+                        if (item.type == "message") {
+                            notificationsDAO.deleteNotification(item.notificationId, item.insertedTime);
+                        }
+
                     }
+                    res.status(200).send({ status: 200, message: "Message notifications deleted successfully" });
                 })
                 .catch((err) => {
                     console.log(err);
